feat(edge/devices): hide untrusted endpoints
